Create PALM_commands_structural+functional
This is an example of the singularity container used to run the general linear models in FSL's Permutation Analysis of Linear Models (PALM). These commands were put together in a txt file and then run using a driver script. These commands are identical for all structural and functional analyses using PALM.